LibWeb: Remove remaining WebAssemblyObject data to the namespace file
